backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
torch
torchvision
numpy>=1.24.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
google-cloud-aiplatform>=1.38.0

# Universal Fake Detect (CLIP-based deepfake detector)
git+https://github.com/openai/CLIP.git
ftfy>=6.1.0
regex>=2023.0.0

# EfficientNet + Grad-CAM (legacy, kept for ensemble option)
efficientnet-pytorch>=0.7.1
grad-cam>=1.5.0

# GenConViT dependencies
face-recognition>=1.3.0
dlib>=19.24.0
albumentations>=1.3.0
decord>=0.6.0
timm==0.6.5
tqdm>=4.65.0
PyYAML>=6.0

transformers
librosa
soundfile
facenet-pytorch

# Video Detection Dependencies
ultralytics
scipy
torchaudio
opencv-python

